+++ .gitignore
@@ -1,3 +1,4 @@
 *.pyc
 actsims.egg-info/
 *.png
+inputParams/paths.yml
