+++ actsims/powtools.py
