+++ inputParams/paths_example.yml
@@ -1,0 +1,4 @@
+
+
+map_root: "/home/msyriac/data/act/maps/mr3/"
+mask_root: "/home/msyriac/data/act/maps/steve/"
